Implement LFA feedback on branches
