+++ Server_db/helper/schema/schema.js
@@ -2,9 +2,9 @@ import Joi from 'joi';
 
 const schema = {
   signup: Joi.object().keys({
-    firstName: Joi.string().trim().regex(/[A-Za-z]/).min(3)
+    firstName: Joi.string().regex(/[A-Za-z]/).min(3)
       .required(),
-    lastName: Joi.string().regex(/[A-Za-z]/).trim().min(3)
+    lastName: Joi.string().regex(/[A-Za-z]/).min(3)
       .required(),
     email: Joi.string().email().lowercase().required(),
     password: Joi.string().required(),
@@ -16,13 +16,13 @@ const schema = {
   }),
 
   account: Joi.object().keys({
-    type: Joi.string().trim().valid('savings', 'current')
+    type: Joi.string().valid('savings', 'current')
       .required(),
     openingBalance: Joi.number().required(),
   }),
 
   setAccount: Joi.object().keys({
-    status: Joi.string().trim().valid('active', 'dormant')
+    status: Joi.string().valid('active', 'dormant')
       .required(),
     accountNumber: Joi.number().integer().required().min(10),
   }),
