+++ Server_db/router/router.js
@@ -1,27 +1,28 @@
 /* eslint-disable linebreak-style */
 import express from 'express';
-import Controller from '../controllers/controller';
+import AccountController from '../controllers/accountcontroller';
+import TransactionController from '../controllers/transactioncontroller';
 import Validate from '../middleware/validator';
-import User from '../controllers/usercontroller';
+import UserController from '../controllers/usercontroller';
 import Authenticator from '../middleware/authenthicator';
 
 // the sub app using url versioning
 const router = express();
 
-router.get('/', Controller.welcome);
-router.get('/accounts/:accountNumber/transactions', Authenticator.user, Validate.accountNumber, Controller.getAllTransactions);
-router.get('/transactions/:id', Authenticator.user, Validate.id, Controller.getTransaction);
-router.get('/user/:email/accounts', Authenticator.user, Validate.email, Controller.getAllUserAccounts);
-router.get('/accounts/:accountNumber', Authenticator.user, Validate.accountNumber, Controller.getAccountDetails);
+router.get('/', UserController.welcome);
 
-router.post('/auth/signup', Validate.Signup, User.signup);
-router.post('/auth/signin', Validate.Signin, User.signin);
+router.post('/auth/signup', Validate.Signup, UserController.signup);
+router.post('/auth/signin', Validate.Signin, UserController.signin);
 
-router.post('/accounts', Authenticator.user, Validate.createAccount, Controller.creatAccount);
-router.post('/transactions/:accountNumber/debit', Authenticator.user, Validate.transaction, Controller.transactions);
-router.post('/transactions/:accountNumber/credit', Authenticator.user, Validate.transaction, Controller.transactions);
+router.get('/accounts/:accountNumber', Authenticator.user, Validate.accountNumber, AccountController.getAccountDetails);
+router.get('/user/:email/accounts', Authenticator.user, Validate.email, AccountController.getAllUserAccounts);
+router.post('/accounts', Authenticator.user, Authenticator.isClient, Validate.createAccount, AccountController.createAccount);
+router.patch('/account/:accountNumber', Authenticator.user, Authenticator.isStaff, Validate.setAccount, AccountController.updateAccount);
+router.delete('/accounts/:accountNumber', Authenticator.user, Authenticator.isStaff, Validate.accountNumber, AccountController.deleteAccount);
 
-router.patch('/account/:accountNumber', Authenticator.user, Validate.setAccount, Controller.updateAccount);
+router.get('/accounts/:accountNumber/transactions', Authenticator.user, Validate.accountNumber, TransactionController.getSingleAccountTransactions);
+router.get('/transactions/:id', Authenticator.user, Validate.id, TransactionController.getSingleTransactionById);
+router.post('/transactions/:accountNumber/debit', Authenticator.user, Validate.transaction, TransactionController.transactions);
+router.post('/transactions/:accountNumber/credit', Authenticator.user, Validate.transaction, TransactionController.transactions);
 
-router.delete('/accounts/:accountNumber', Authenticator.user, Validate.accountNumber, Controller.deleteAccount);
 export default router;
